Fixed Loading Timetable issue due to cookie expiry date
I noticed that there was a feature to load any saved
timetables but somehow couldnt locate the button.

It turns out that this is due to the expiry date of the
cookie being used to save the timetable is set to 1 Jan 2010,
which is a date that's already past.

I changed the date to 1 Jan 3010 to resolve the issue.